requirements.txt
streamlit
pypdf
reportlab
pillow
selenium
webdriver-manager
requests
PyPDF2